src/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
mongoengine
google-generativeai
reportlab